prepare FFI. > testcase.input is undefined. > testcase.output is undefined.
