--- .github/workflows/push-to-gitea.yml
@@ -1,29 +1,0 @@
-## gitea に同期する用の actions
-name: main
-
-on:
-  push:
-    branches:
-      - "main"
-
-env:
-  GITEA_SSH_URL: "ssh://git@git.trap.jp:2200/Hackathon22-Winter-03/backend.git"
-
-jobs:
-  sync:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Set config
-        run: |
-          echo "${{ secrets.GITEA_TOKEN }}" > ~/deploy_key.pem
-          chmod 600 ~/deploy_key.pem
-          git config --local user.email "github@actions"
-          git config --local user.name "github"
-      - name: Push changes
-        env:
-          GIT_SSH_COMMAND: ssh -i ~/deploy_key.pem -o StrictHostKeyChecking=no -F /dev/null
-        run: |
-          git fetch --unshallow origin
-          git push -f ${{ env.GITEA_SSH_URL }} main:master
