+++ Makefile
@@ -1,7 +1,7 @@
 ROOT_DIR := $(dir $(realpath $(lastword $(MAKEFILE_LIST))))
 
 build:
-    cd lang && cargo build --release
-    cp lang/target/release/librustaceanize.dylib lib/
-    echo 'ROOT_DIR is $(ROOT_DIR)'
-    go build -ldflags="-r $(ROOT_DIR)lib" main.go
+	cd lang && cargo build --release
+	cp lang/target/release/liblang.so lang/
+	echo 'ROOT_DIR is $(ROOT_DIR)'
+	go build -ldflags="-r $(ROOT_DIR)lang" main.go
