+++ lang/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2021"
 crate-type = ["cdylib"]
 
 [dependencies]
+libc = "0.2.2"
